introduce a heterogenoues SoA for Hits
